Restructure Legend
- use private, hooks and utils dirs
- Update exports
