+++ projects/js-packages/charts/src/components/legend/index.ts
@@ -1,5 +1,4 @@
 export { Legend } from './legend';
-export { BaseLegend } from './base-legend';
-export { useChartLegendItems } from './use-chart-legend-items';
+export { useChartLegendItems } from './hooks/use-chart-legend-items';
 export type { LegendProps, BaseLegendProps } from './types';
-export type { ChartLegendOptions } from './use-chart-legend-items';
+export type { ChartLegendOptions } from './hooks/use-chart-legend-items';
