+++ projects/js-packages/charts/src/components/legend/private/legend.module.scss
