+++ projects/js-packages/charts/src/components/legend/utils/index.ts
@@ -1,0 +1,3 @@
+export { valueOrIdentity, valueOrIdentityString } from './value-or-identity';
+export { labelTransformFactory } from './label-transform-factory';
+export type { ValueOrIdentity } from './value-or-identity';
