+++ projects/js-packages/charts/src/components/pie-semi-circle-chart/pie-semi-circle-chart.tsx
@@ -15,8 +15,7 @@ import {
 } from '../../providers/chart-context';
 import { attachSubComponents } from '../../utils';
 import { ChartSVG, ChartHTML, useChartChildren } from '../chart-composition';
-import { Legend } from '../legend';
-import { useChartLegendItems } from '../legend/use-chart-legend-items';
+import { Legend, useChartLegendItems } from '../legend';
 import { SingleChartContext } from '../single-chart-context';
 import { BaseTooltip } from '../tooltip';
 import { withResponsive } from '../with-responsive';
